Disable telemetry by default
Currently, telemetry features default to enabled, which can cause problems for
deployment models such as colocated where the user has not met the prerequisites.

It is, therefore, more conservative to default to disabled and require the user
to explicitly opt-in via inventory configuration.  This requirement will result in a consistent
configuration across deployment and upgrade since the inventory will be common to both modes.

N.B.  The most significant implication here is that the embedded model is now changing
behavior, requiring the user to opt-in here as well whereas they previously did
not.

Signed-off-by: Greg Haskins <greg@manetu.com>