Add support for "custom" persistence driver
Temporal core allows for the registration and configuration of custom data stores.
However, there is a disparity in how the helm chart translates persistence
configuration between helm values and the resulting configmap.

As an example, the temporal config for a custom driver for the default store may look
as so:

```
persistence:
  defaultStore: my-custom-store
  visibilityStore: es-visibility
  numHistoryShards: 4
  datastores:
    my-custom-store:
      customDatastore:
        name: "my-custom"
        options:
          hosts: "mycustom.example.com"
          keyspace: "temporal"
```

Translating to helm values would look like

```
server:
  config:
    persistence:
      defaultStore: my-custom-store
      additionalStores:
        my-custom-store:
          customDatastore:
            name: "my-custom"
            options:
              hosts: "mycustom.example.com"
              keyspace: "temporal"
```

This gets us most of the way there, but there are a few problems:

1. The charts always assume defaultStore = "default" and thus have opinions
   about subordinate structures such as default.driver.  This results in objects
   such as Secrets and Jobs being incorrectly emitted for default.driver = "cassandra" even
   though we are trying to specify an alternate store.  This logic should probably
   be $defaultStore.driver and $visibilityStore.driver instead.
2. The additionalStores configuration does not lend itself to first-class
   representation in conditional statements, even though it can be used to
   emit the resulting configmap correctly.

One could argue that we should clean up the helm representation to more closely align
with the temporal-server yaml, but this is a bigger task (and a breaking change).  Instead,
this patch ties into the existing model of using $store.driver = {cassandra, sql, elasticsearch}
and adds a 4th first-class type "custom."

The user may configure this as so:

```
server:
  config:
    persistence:
      default:
        driver: custom
      visibility:
        driver: custom
```

And in doing so, we instruct the various templates to elide the objects (Jobs, Secrets, etc.)
related to that store, leaving it to the user to supply the rest of the required configuration
using a combination of additionalStores, additionalEnv, additionalVolumeMounts, etc., to inject
configuration and secrets.  They would also, of course, be responsible for any DDL setup or
migration.

Signed-off-by: Greg Haskins <greg@manetu.com>