Up the temporal RPS/QPS settings
We have long observed a phenomenon where temporal performance is increased by
over-provisioning replicas of the temporal pods, even though they were not maxing
out CPU per se.  It turns out that there are per-pod limits for requests-per-second
and queries-per-second, and we are often hitting them when operating at the edge
as evident from the resource-exhaustion metrics.

In the past, we have added more pods and observed increased performance.  We should
be able to achieve the same effect by increasing the RPS/QPS limits.  We may even
want to set them artifcially high and let the node/persistence layer naturally
be the bottleneck, but this needs more research.  For now, we simply double the
capacity.

Signed-off-by: Greg Haskins <greg@manetu.com>