Rename Istio ingress-gw to ingressgateway
This better aligns with how default installations of Istio
name the ingress resource.

We unfortunately must coordinate with ci-tools because it has
a coupled relationship.  Therefore, we also disable ITF and bump
the version to v1.1.0 to perform the codependent dance.

The ITF will be re-enabled after ci-tools is updated

Signed-off-by: Greg Haskins <greg@manetu.com>